# openapi/kayan_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kayan Orphan Care — AI Agent Platform (Mock)",
    "description": "Prototype backend simulating جمعية كيان للايتام beneficiary management. Multi-channel (Voice/SIP + WhatsApp), Arabic-first. Each route is an agent tool.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/programs/{program_id}/request-types": {
      "get": {
        "tags": [
          "5 · طلبات الدعم | Support Requests"
        ],
        "summary": "List request types under a program",
        "description": "The projects/requests available under a program, with the approval ceiling and whether the support is recurring monthly or one-off.",
        "operationId": "program_requests_programs__program_id__request_types_get",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Program Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/request-types/search": {
      "get": {
        "tags": [
          "5 · طلبات الدعم | Support Requests"
        ],
        "summary": "Find the right request type from free text",
        "description": "Maps what the beneficiary said (e.g. 'محتاج مساعدة في الايجار') to the correct request type and program. Use this before creating a request so the case lands in the right program.",
        "operationId": "search_request_types_request_types_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "examples": [
                "ايجار"
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/support-requests": {
      "post": {
        "tags": [
          "5 · طلبات الدعم | Support Requests"
        ],
        "summary": "Submit a support request (طلب دعم)",
        "description": "Creates a support request under a program. Requires an APPROVED beneficiary file — returns 409 otherwise, matching the guide ('after your file is complete and approved'). Amounts above the request ceiling are rejected.",
        "operationId": "create_request_support_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/support-requests/{request_id}": {
      "get": {
        "tags": [
          "5 · طلبات الدعم | Support Requests"
        ],
        "summary": "Get a support request with its casework and decision",
        "description": "Full request detail including the case study steps and the committee decision if one has been issued.",
        "operationId": "get_request_support_requests__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiary/{beneficiary_id}/support-requests": {
      "get": {
        "tags": [
          "5 · طلبات الدعم | Support Requests"
        ],
        "summary": "List a beneficiary's support requests",
        "description": "All requests raised by a beneficiary with their stage and decision.",
        "operationId": "list_requests_beneficiary__beneficiary_id__support_requests_get",
        "parameters": [
          {
            "name": "beneficiary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beneficiary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/support-requests/{request_id}/add-detail": {
      "patch": {
        "tags": [
          "5 · طلبات الدعم | Support Requests"
        ],
        "summary": "Append more detail to a request",
        "description": "Use when the case worker or bot asks the beneficiary to explain further. Appends to وصف الحالة rather than overwriting it, preserving the history.",
        "operationId": "add_detail_support_requests__request_id__add_detail_patch",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDetailIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/support-requests/{request_id}/open-case": {
      "post": {
        "tags": [
          "6 · دراسة الحالة واللجنة | Casework & Committee"
        ],
        "summary": "Open a case study for a request",
        "description": "Starts دراسة الحالة and moves the request to 'under_study'. Assigns a social researcher.",
        "operationId": "open_case_support_requests__request_id__open_case_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "researcher_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "examples": [
                "STF-04"
              ],
              "default": "STF-04",
              "title": "Researcher Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/schedule-step": {
      "post": {
        "tags": [
          "6 · دراسة الحالة واللجنة | Casework & Committee"
        ],
        "summary": "Schedule a case study step (visit / interview / assessment)",
        "description": "Schedules a field visit, office or online interview, or psychological assessment, and notifies the beneficiary on WhatsApp.",
        "operationId": "schedule_step_cases__case_id__schedule_step_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleStepIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/record-findings": {
      "post": {
        "tags": [
          "6 · دراسة الحالة واللجنة | Casework & Committee"
        ],
        "summary": "Record findings for a completed step",
        "description": "Marks a case step complete and stores the researcher's findings.",
        "operationId": "record_findings_cases__case_id__record_findings_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/submit-to-committee": {
      "post": {
        "tags": [
          "6 · دراسة الحالة واللجنة | Casework & Committee"
        ],
        "summary": "Submit the case to the specialized committee",
        "description": "Presents the case to اللجنة المختصة with the researcher's recommendation. Requires at least one completed step (409 otherwise).",
        "operationId": "submit_committee_cases__case_id__submit_to_committee_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCommitteeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/support-requests/{request_id}/decision": {
      "post": {
        "tags": [
          "6 · دراسة الحالة واللجنة | Casework & Committee"
        ],
        "summary": "Record the committee decision and notify the beneficiary",
        "description": "Records قبول الطلب / طلب استكمال مستندات / الاعتذار, then notifies the beneficiary on BOTH WhatsApp and SMS as the guide specifies. Accepted requests become eligible for enrollment and disbursement.",
        "operationId": "record_decision_support_requests__request_id__decision_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddDetailIn": {
        "properties": {
          "additional_detail_ar": {
            "type": "string",
            "minLength": 5,
            "title": "Additional Detail Ar",
            "examples": [
              "صدر انذار اخلاء من المالك بتاريخ 1 يوليو"
            ]
          }
        },
        "type": "object",
        "required": [
          "additional_detail_ar"
        ],
        "title": "AddDetailIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleStepIn": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id",
            "description": "CS-DESK | CS-ONLINE | CS-FIELD | CS-PSYCH",
            "examples": [
              "CS-FIELD"
            ]
          },
          "scheduled_at": {
            "type": "string",
            "title": "Scheduled At",
            "examples": [
              "2026-08-02T10:00:00Z"
            ]
          },
          "assigned_staff_id": {
            "type": "string",
            "title": "Assigned Staff Id",
            "default": "STF-04",
            "examples": [
              "STF-04"
            ]
          }
        },
        "type": "object",
        "required": [
          "step_id",
          "scheduled_at"
        ],
        "title": "ScheduleStepIn"
      },
...
    }
  }
}